+++ .changeset/witty-squids-dance.md
@@ -1,0 +1,8 @@
+---
+'@pizzajsdev/react-router-hono': minor
+'@pizzajsdev/app-router-fs': minor
+'@pizzajsdev/utils': minor
+'@pizzajsdev/i18n': minor
+---
+
+upgrade dependencies and migrate to tsdown
