+++ package.json
@@ -30,10 +30,14 @@
     "@changesets/cli": "^2.29.5",
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
+    "@oxc-project/runtime": "^0.80.0",
+    "@types/node": "^24.2.0",
     "husky": "^9.1.7",
     "madge": "^8.0.0",
     "prettier": "^3.6.2",
+    "publint": "^0.3.12",
     "sort-package-json": "^3.4.0",
+    "tsdown": "^0.13.3",
     "typescript": "^5.9.2",
     "vitest": "^3.2.4"
   },
