+++ packages/app-router-fs/package.json
@@ -31,16 +31,17 @@
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist && tsup --clean",
+    "build": "rm -rf dist && tsdown --clean",
     "postbuild": "publint",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
+    "@oxc-project/runtime": "^0.80.0",
     "@react-router/dev": "^7.7.1",
     "@types/node": "^24.2.0",
     "publint": "^0.3.12",
-    "tsup": "^8.5.0",
+    "tsdown": "^0.13.3",
     "typescript": "^5.9.2"
   }
 }
