+++ packages/i18n/package.json
@@ -25,22 +25,23 @@
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist && tsup --clean",
+    "build": "rm -rf dist && tsdown --clean",
     "postbuild": "publint",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@pizzajsdev/utils": "workspace:*"
   },
   "devDependencies": {
+    "@oxc-project/runtime": "^0.80.0",
     "@types/node": "^24.2.0",
     "@types/react": "^19.1.9",
     "@types/react-dom": "^19.1.7",
     "publint": "^0.3.12",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "tsup": "^8.5.0",
+    "tsdown": "^0.13.3",
     "typescript": "^5.9.2"
   }
 }
