+++ packages/react-router-hono/package.json
@@ -25,9 +25,9 @@
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist && tsup --clean",
+    "build": "rm -rf dist && tsdown --clean",
     "postbuild": "publint",
-    "dev": "tsup --watch",
+    "dev": "tsdown --watch",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
@@ -39,6 +39,7 @@
     "vite": "^7.0.6"
   },
   "devDependencies": {
+    "@oxc-project/runtime": "^0.80.0",
     "@react-router/dev": "^7.7.1",
     "@react-router/node": "^7.7.1",
     "@types/node": "^24.2.0",
@@ -49,7 +50,7 @@
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-router": "^7.7.1",
-    "tsup": "^8.5.0",
+    "tsdown": "^0.13.3",
     "typescript": "^5.9.2"
   }
 }
