+++ packages/utils/tsdown.config.ts
@@ -1,4 +1,4 @@
-import { type Options, defineConfig } from 'tsup'
+import { type Options, defineConfig } from 'tsdown'
 
 const config: Options = {
   entry: ['./src/**/!(*.test).ts'],
@@ -11,10 +11,11 @@ const config: Options = {
   bundle: false,
   dts: true,
   sourcemap: true,
-  splitting: true,
   treeshake: true,
   skipNodeModulesBundle: true,
   external: ['node_modules'],
+  // @ts-expect-error - splitting is not supported by tsdown yet, but it's WIP
+  splitting: true,
 }
 
 export default defineConfig([config])
